+++ CHANGELOG.md
@@ -46,6 +46,10 @@
     FIXES [#6635](https://github.com/microsoft/Microsoft365DSC/issues/6635)
 * VivaEngagementRoleMember
   * Fixed issue loading the required modules to export this resource
+* M365DSCReverse
+  * Fixed an issue where having resources specified in both include and exclude
+    during export would result in an exception.
+    FIXES [#6630](https://github.com/microsoft/Microsoft365DSC/issues/6630)
 * M365DSCUtil
   * Added custom post processing to `Test-M365DSCTargetResource`.
   * Moved resource initialization into dedicated function.